Some refactoring.
Remove the shutdown command since it's only useful when wait is used
with --no-shutdown which is rarely done and it's confusing to have an
explicit shutdown when wait implicitly shuts down.

Add doc strings to command line arguments to provide a better --help
message.

Refactor the private module structure and how request and responses are
sent and received.